requirements.txt
autopep8==2.3.2
blinker==1.9.0
click==8.2.0
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
pycodestyle==2.13.0
pytesseract==0.3.13
python-docx==1.1.2
typing_extensions==4.13.2
Werkzeug==3.0.6
